refactored discoverylistener to it's own class file, update relevant javadocs
