+++ integration/src/androidTest/java/com/idevicesinc/sweetblue/NotifyTest.java
@@ -34,7 +34,7 @@ public class NotifyTest extends BaseTester<MainActivity>
     public void isNotifyEnabledTest() throws Exception
     {
         final Semaphore s = new Semaphore(0);
-        mgr.setListener_Discovery(new BleManager.DiscoveryListener()
+        mgr.setListener_Discovery(new DiscoveryListener()
         {
             @Override public void onEvent(DiscoveryEvent e)
             {
